requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiogram==2.23.1
python-dotenv==0.21.0
